enable email notifications when subscribing to company topic
and move topic handling to topics module

NHUB-370
