+++ newsroom/topics/views.py
@@ -1,3 +1,5 @@
+import flask
+
 from bson import ObjectId
 from superdesk import get_resource_service
 from flask import json, jsonify, abort, current_app as app, url_for
@@ -17,10 +19,39 @@ from newsroom.notifications import (
 )
 
 
-@blueprint.route("/users/<user_id>/topics", methods=["GET"])
+@blueprint.route("/users/<_id>/topics", methods=["GET"])
 @login_required
-def get_user_topics(user_id):
-    return jsonify(_get_user_topics(user_id)), 200
+def get_topics(_id):
+    """Returns list of followed topics of given user"""
+    if flask.session["user"] != str(_id):
+        flask.abort(403)
+    return jsonify({"_items": _get_user_topics(_id)}), 200
+
+
+@blueprint.route("/users/<_id>/topics", methods=["POST"])
+@login_required
+def post_topic(_id):
+    """Creates a user topic"""
+    user = get_user()
+    if str(user["_id"]) != str(_id):
+        flask.abort(403)
+
+    topic = get_json_or_400()
+    topic["user"] = user["_id"]
+    topic["company"] = user.get("company")
+
+    for subscriber in topic.get("subscribers") or []:
+        subscriber["user_id"] = ObjectId(subscriber["user_id"])
+
+    ids = get_resource_service("topics").post([topic])
+
+    auto_enable_user_emails(topic, {}, user)
+
+    if topic.get("is_global"):
+        push_company_notification("topic_created", user_id=str(user["_id"]))
+    else:
+        push_user_notification("topic_created")
+    return jsonify({"success": True, "_id": ids[0]}), 201
 
 
 @blueprint.route("/topics/my_topics", methods=["GET"])
